+++ app/kumactl/cmd/get/get.go
@@ -16,6 +16,14 @@ type getContext struct {
 	}
 }
 
+type listContext struct {
+	*getContext
+	args struct {
+		size   int
+		offset string
+	}
+}
+
 func NewGetCmd(pctx *kumactl_cmd.RootContext) *cobra.Command {
 	ctx := &getContext{RootContext: pctx}
 	cmd := &cobra.Command{
@@ -26,15 +34,17 @@ func NewGetCmd(pctx *kumactl_cmd.RootContext) *cobra.Command {
 	// flags
 	cmd.PersistentFlags().StringVarP(&ctx.args.outputFormat, "output", "o", string(output.TableFormat), kuma_cmd.UsageOptions("output format", output.TableFormat, output.YAMLFormat, output.JSONFormat))
 	// sub-commands
-	cmd.AddCommand(newGetMeshesCmd(ctx))
-	cmd.AddCommand(newGetDataplanesCmd(ctx))
-	cmd.AddCommand(newGetHealthChecksCmd(ctx))
-	cmd.AddCommand(newGetProxyTemplatesCmd(ctx))
-	cmd.AddCommand(newGetTrafficPermissionsCmd(ctx))
-	cmd.AddCommand(newGetTrafficRoutesCmd(ctx))
-	cmd.AddCommand(newGetTrafficLogsCmd(ctx))
-	cmd.AddCommand(newGetTrafficTracesCmd(ctx))
-	cmd.AddCommand(newGetFaultInjectionsCmd(ctx))
+	listCtx := &listContext{getContext: ctx}
+	cmd.AddCommand(withPaginationArgs(newGetMeshesCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetDataplanesCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetHealthChecksCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetProxyTemplatesCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetTrafficPermissionsCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetTrafficRoutesCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetTrafficLogsCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetTrafficTracesCmd(listCtx), listCtx))
+	cmd.AddCommand(withPaginationArgs(newGetFaultInjectionsCmd(listCtx), listCtx))
+
 	cmd.AddCommand(newGetFaultInjectionCmd(ctx))
 	cmd.AddCommand(newGetMeshCmd(ctx))
 	cmd.AddCommand(newGetDataplaneCmd(ctx))
@@ -44,5 +54,11 @@ func NewGetCmd(pctx *kumactl_cmd.RootContext) *cobra.Command {
 	cmd.AddCommand(newGetTrafficPermissionCmd(ctx))
 	cmd.AddCommand(newGetTrafficRouteCmd(ctx))
 	cmd.AddCommand(newGetTrafficTraceCmd(ctx))
+	return cmd
+}
+
+func withPaginationArgs(cmd *cobra.Command, ctx *listContext) *cobra.Command {
+	cmd.PersistentFlags().IntVarP(&ctx.args.size, "size", "", 0, "maximum number of elements to return")
+	cmd.PersistentFlags().StringVarP(&ctx.args.offset, "offset", "", "", "the offset that indicates starting element of the resources list to retrieve")
 	return cmd
 }