+++ app/kumactl/cmd/get/get_fault_injection.go
@@ -35,7 +35,9 @@ func newGetFaultInjectionCmd(pctx *getContext) *cobra.Command {
 				}
 				return errors.Wrapf(err, "failed to get mesh %s", currentMesh)
 			}
-			faultInjections := []*mesh.FaultInjectionResource{faultInjection}
+			faultInjections := &mesh.FaultInjectionResourceList{
+				Items: []*mesh.FaultInjectionResource{faultInjection},
+			}
 			switch format := output.Format(pctx.args.outputFormat); format {
 			case output.TableFormat:
 				return printFaultInjections(faultInjections, cmd.OutOrStdout())