+++ app/kumactl/cmd/get/get_fault_injections.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"io"
 
+	"github.com/Kong/kuma/app/kumactl/pkg/output/table"
+
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -14,7 +16,7 @@ import (
 	core_store "github.com/Kong/kuma/pkg/core/resources/store"
 )
 
-func newGetFaultInjectionsCmd(pctx *getContext) *cobra.Command {
+func newGetFaultInjectionsCmd(pctx *listContext) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "fault-injections",
 		Short: "Show FaultInjections",
@@ -26,13 +28,13 @@ func newGetFaultInjectionsCmd(pctx *getContext) *cobra.Command {
 			}
 
 			faultInjections := mesh.FaultInjectionResourceList{}
-			if err := rs.List(context.Background(), &faultInjections, core_store.ListByMesh(pctx.CurrentMesh())); err != nil {
+			if err := rs.List(context.Background(), &faultInjections, core_store.ListByMesh(pctx.CurrentMesh()), core_store.ListByPage(pctx.args.size, pctx.args.offset)); err != nil {
 				return errors.Wrapf(err, "failed to list FaultInjection")
 			}
 
-			switch format := output.Format(pctx.args.outputFormat); format {
+			switch format := output.Format(pctx.getContext.args.outputFormat); format {
 			case output.TableFormat:
-				return printFaultInjections(faultInjections.Items, cmd.OutOrStdout())
+				return printFaultInjections(&faultInjections, cmd.OutOrStdout())
 			default:
 				printer, err := printers.NewGenericPrinter(format)
 				if err != nil {
@@ -45,17 +47,17 @@ func newGetFaultInjectionsCmd(pctx *getContext) *cobra.Command {
 	return cmd
 }
 
-func printFaultInjections(faultInjections []*mesh.FaultInjectionResource, out io.Writer) error {
+func printFaultInjections(faultInjections *mesh.FaultInjectionResourceList, out io.Writer) error {
 	data := printers.Table{
 		Headers: []string{"MESH", "NAME"},
 		NextRow: func() func() []string {
 			i := 0
 			return func() []string {
 				defer func() { i++ }()
-				if len(faultInjections) <= i {
+				if len(faultInjections.Items) <= i {
 					return nil
 				}
-				faultInjection := faultInjections[i]
+				faultInjection := faultInjections.Items[i]
 
 				return []string{
 					faultInjection.GetMeta().GetMesh(), // MESH
@@ -63,6 +65,7 @@ func printFaultInjections(faultInjections []*mesh.FaultInjectionResource, out io
 				}
 			}
 		}(),
+		Footer: table.PaginationFooter(faultInjections),
 	}
 	return printers.NewTablePrinter().Print(data, out)
 }