+++ app/kumactl/cmd/get/get_proxytemplate.go
@@ -35,7 +35,9 @@ func newGetProxyTemplateCmd(pctx *getContext) *cobra.Command {
 				}
 				return errors.Wrapf(err, "failed to get mesh %s", currentMesh)
 			}
-			proxyTemplates := []*mesh.ProxyTemplateResource{proxyTemplate}
+			proxyTemplates := &mesh.ProxyTemplateResourceList{
+				Items: []*mesh.ProxyTemplateResource{proxyTemplate},
+			}
 			switch format := output.Format(pctx.args.outputFormat); format {
 			case output.TableFormat:
 				return printProxyTemplates(proxyTemplates, cmd.OutOrStdout())