+++ app/kumactl/cmd/get/get_proxytemplates.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"io"
 
+	"github.com/Kong/kuma/app/kumactl/pkg/output/table"
+
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -14,7 +16,7 @@ import (
 	core_store "github.com/Kong/kuma/pkg/core/resources/store"
 )
 
-func newGetProxyTemplatesCmd(pctx *getContext) *cobra.Command {
+func newGetProxyTemplatesCmd(pctx *listContext) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "proxytemplates",
 		Short: "Show ProxyTemplates",
@@ -26,13 +28,13 @@ func newGetProxyTemplatesCmd(pctx *getContext) *cobra.Command {
 			}
 
 			proxyTemplates := &mesh_core.ProxyTemplateResourceList{}
-			if err := rs.List(context.Background(), proxyTemplates, core_store.ListByMesh(pctx.CurrentMesh())); err != nil {
+			if err := rs.List(context.Background(), proxyTemplates, core_store.ListByMesh(pctx.CurrentMesh()), core_store.ListByPage(pctx.args.size, pctx.args.offset)); err != nil {
 				return errors.Wrapf(err, "failed to list ProxyTemplates")
 			}
 
-			switch format := output.Format(pctx.args.outputFormat); format {
+			switch format := output.Format(pctx.getContext.args.outputFormat); format {
 			case output.TableFormat:
-				return printProxyTemplates(proxyTemplates.Items, cmd.OutOrStdout())
+				return printProxyTemplates(proxyTemplates, cmd.OutOrStdout())
 			default:
 				printer, err := printers.NewGenericPrinter(format)
 				if err != nil {
@@ -45,17 +47,17 @@ func newGetProxyTemplatesCmd(pctx *getContext) *cobra.Command {
 	return cmd
 }
 
-func printProxyTemplates(proxyTemplates []*mesh_core.ProxyTemplateResource, out io.Writer) error {
+func printProxyTemplates(proxyTemplates *mesh_core.ProxyTemplateResourceList, out io.Writer) error {
 	data := printers.Table{
 		Headers: []string{"MESH", "NAME"},
 		NextRow: func() func() []string {
 			i := 0
 			return func() []string {
 				defer func() { i++ }()
-				if len(proxyTemplates) <= i {
+				if len(proxyTemplates.Items) <= i {
 					return nil
 				}
-				proxyTemplate := proxyTemplates[i]
+				proxyTemplate := proxyTemplates.Items[i]
 
 				return []string{
 					proxyTemplate.GetMeta().GetMesh(), // MESH
@@ -63,6 +65,7 @@ func printProxyTemplates(proxyTemplates []*mesh_core.ProxyTemplateResource, out 
 				}
 			}
 		}(),
+		Footer: table.PaginationFooter(proxyTemplates),
 	}
 	return printers.NewTablePrinter().Print(data, out)
 }