+++ app/kumactl/cmd/get/get_traffic_logs_test.go
@@ -116,6 +116,7 @@ var _ = Describe("kumactl get traffic-logs", func() {
 
 		type testCase struct {
 			outputFormat string
+			pagination   string
 			goldenFile   string
 			matcher      func(interface{}) gomega_types.GomegaMatcher
 		}
@@ -125,7 +126,7 @@ var _ = Describe("kumactl get traffic-logs", func() {
 				// given
 				rootCmd.SetArgs(append([]string{
 					"--config-file", filepath.Join("..", "testdata", "sample-kumactl.config.yaml"),
-					"get", "traffic-logs"}, given.outputFormat))
+					"get", "traffic-logs"}, given.outputFormat, given.pagination))
 
 				// when
 				err := rootCmd.Execute()
@@ -149,6 +150,14 @@ var _ = Describe("kumactl get traffic-logs", func() {
 			Entry("should support Table output explicitly", testCase{
 				outputFormat: "-otable",
 				goldenFile:   "get-traffic-logs.golden.txt",
+				matcher: func(expected interface{}) gomega_types.GomegaMatcher {
+					return WithTransform(strings.TrimSpace, Equal(strings.TrimSpace(string(expected.([]byte)))))
+				},
+			}),
+			Entry("should support pagination", testCase{
+				outputFormat: "-otable",
+				pagination:   "--size=1",
+				goldenFile:   "get-traffic-logs.pagination.golden.txt",
 				matcher: func(expected interface{}) gomega_types.GomegaMatcher {
 					return WithTransform(strings.TrimSpace, Equal(strings.TrimSpace(string(expected.([]byte)))))
 				},