+++ app/kumactl/cmd/get/get_traffic_trace.go
@@ -33,7 +33,9 @@ func newGetTrafficTraceCmd(pctx *getContext) *cobra.Command {
 				}
 				return errors.Wrapf(err, "failed to get mesh %s", currentMesh)
 			}
-			trafficTraces := []*mesh.TrafficTraceResource{trafficTrace}
+			trafficTraces := &mesh.TrafficTraceResourceList{
+				Items: []*mesh.TrafficTraceResource{trafficTrace},
+			}
 			switch format := output.Format(pctx.args.outputFormat); format {
 			case output.TableFormat:
 				return printTrafficTraces(trafficTraces, cmd.OutOrStdout())