+++ pkg/core/resources/model/rest/resource_test.go
@@ -81,7 +81,8 @@ var _ = Describe("Rest Resource", func() {
 					"name": "two",
 					"path": "/another"
 				 }
-				]
+				],
+				"next": "http://localhost:5681/meshes/default/traffic-routes?offset=1"
 			}`
 
 				// when
@@ -118,7 +119,9 @@ var _ = Describe("Rest Resource", func() {
 						Spec: &sample_proto.TrafficRoute{
 							Path: "/another",
 						},
-					}))
+					}),
+				)
+				Expect(*rs.Next).To(Equal("http://localhost:5681/meshes/default/traffic-routes?offset=1"))
 			})
 		})
 	})